requirements.txt
streamlit==1.35.0
ultralytics==8.1.1
opencv-python==4.9.0.80
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
python-dotenv==1.0.1
scikit-learn==1.4.2
openai==1.30.1